# docs/api/pipeline-openapi.json
{
  "components": {
    "schemas": {
...
      "EmbeddingBundleSchema": {
        "additionalProperties": false,
        "description": "Durable state of one embedding bundle. Never carries vectors.",
        "properties": {
          "bundle_schema_version": {
            "title": "Bundle Schema Version",
            "type": "string"
          },
          "checksums": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Checksums",
            "type": "object"
          },
          "configuration_fingerprint": {
            "title": "Configuration Fingerprint",
            "type": "string"
          },
          "corpus_version": {
            "title": "Corpus Version",
            "type": "string"
          },
          "dimension": {
            "title": "Dimension",
            "type": "integer"
          },
          "distance_metric": {
            "title": "Distance Metric",
            "type": "string"
          },
          "embedding_bundle_id": {
            "title": "Embedding Bundle Id",
            "type": "string"
          },
          "embedding_profile_id": {
            "title": "Embedding Profile Id",
            "type": "string"
          },
          "links": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Links",
            "type": "object"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "model_revision": {
            "title": "Model Revision",
            "type": "string"
          },
          "normalization": {
            "title": "Normalization",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "readiness_status": {
            "title": "Readiness Status",
            "type": "string"
          },
          "sealed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sealed At"
          },
          "source_chunk_bundle_id": {
            "title": "Source Chunk Bundle Id",
            "type": "string"
          },
          "source_content_fingerprint": {
            "title": "Source Content Fingerprint",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "validation_status": {
            "title": "Validation Status",
            "type": "string"
          },
          "vector_count": {
            "title": "Vector Count",
            "type": "integer"
          },
          "vector_dtype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vector Dtype"
          },
          "vector_shape": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vector Shape"
          }
        },
        "required": [
          "embedding_bundle_id",
          "source_chunk_bundle_id",
          "embedding_profile_id",
          "provider",
          "model",
          "model_revision",
          "dimension",
          "normalization",
          "distance_metric",
          "configuration_fingerprint",
          "corpus_version",
          "bundle_schema_version",
          "source_content_fingerprint",
          "vector_count",
          "status",
          "validation_status",
          "readiness_status"
        ],
        "title": "EmbeddingBundleSchema",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Chatbot SST Pipeline API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/embedding/bundles/{embedding_bundle_id}": {
      "get": {
        "operationId": "get_bundle_api_embedding_bundles__embedding_bundle_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "embedding_bundle_id",
            "required": true,
            "schema": {
              "title": "Embedding Bundle Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbeddingBundleSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Bundle",
        "tags": [
          "embedding"
        ]
      }
    },
    "/api/embedding/bundles/{embedding_bundle_id}/chunks": {
      "get": {
        "operationId": "list_bundle_chunks_api_embedding_bundles__embedding_bundle_id__chunks_get",
        "parameters": [
          {
            "in": "path",
            "name": "embedding_bundle_id",
            "required": true,
            "schema": {
              "title": "Embedding Bundle Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedBundleChunksSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "List Bundle Chunks",
        "tags": [
          "embedding"
        ]
      }
    },
    "/api/embedding/bundles/{embedding_bundle_id}/indexing-readiness": {
      "get": {
        "operationId": "get_bundle_indexing_readiness_api_embedding_bundles__embedding_bundle_id__indexing_readiness_get",
        "parameters": [
          {
            "in": "path",
            "name": "embedding_bundle_id",
            "required": true,
            "schema": {
              "title": "Embedding Bundle Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbeddingIndexingReadinessSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Bundle Indexing Readiness",
        "tags": [
          "embedding"
        ]
      }
    },
    "/api/embedding/bundles/{embedding_bundle_id}/validation": {
      "get": {
        "operationId": "get_bundle_validation_api_embedding_bundles__embedding_bundle_id__validation_get",
        "parameters": [
          {
            "in": "path",
            "name": "embedding_bundle_id",
            "required": true,
            "schema": {
              "title": "Embedding Bundle Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbeddingBundleValidationSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Bundle Validation",
        "tags": [
          "embedding"
        ]
      }
    },
    "/api/embedding/chunk-bundles": {
      "get": {
        "operationId": "list_chunk_bundles_api_embedding_chunk_bundles_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedChunkBundlesSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "List Chunk Bundles",
        "tags": [
          "embedding"
        ]
      }
    },
    "/api/embedding/chunk-bundles/{chunk_bundle_id}/summary": {
      "get": {
        "operationId": "get_chunk_bundle_summary_api_embedding_chunk_bundles__chunk_bundle_id__summary_get",
        "parameters": [
          {
            "in": "path",
            "name": "chunk_bundle_id",
            "required": true,
            "schema": {
              "title": "Chunk Bundle Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChunkBundleSummarySchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Chunk Bundle Summary",
        "tags": [
          "embedding"
        ]
      }
    },
    "/api/embedding/profiles": {
      "get": {
        "operationId": "list_profiles_api_embedding_profiles_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedProfilesSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "List Profiles",
        "tags": [
          "embedding"
        ]
      }
    },
...
    "/api/embedding/runs/{embedding_run_id}": {
      "get": {
        "operationId": "get_run_api_embedding_runs__embedding_run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "embedding_run_id",
            "required": true,
            "schema": {
              "title": "Embedding Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbeddingRunSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Run",
        "tags": [
          "embedding"
        ]
      }
    },
...
  }
}